レスポンスとリクエストを分離 #20
